abci: make tendermint example+test clients manage a mutex
